requirements.txt
flask
flask-mysqldb
Werkzeug
mysqlclient
python-dotenv